refactor: KeyboardManager로 중복코드 간소화, TimeZone을 current로 받아오도록 수정
