+++ Diary/Controller/NewDiaryViewController.swift
@@ -15,12 +15,14 @@ final class NewDiaryViewController: UIViewController {
         return view
     }()
     
+    private var keyboardManager: KeyboardManager?
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         configureNavigation()
         configureUI()
         configureLayout()
-        setUpKeyboardEvent()
+        setUpKeyboard()
     }
 
     private func configureNavigation() {
@@ -42,26 +44,7 @@ final class NewDiaryViewController: UIViewController {
         ])
     }
     
-    private func setUpKeyboardEvent() {
-        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillShow), name: UIResponder.keyboardWillShowNotification, object: nil)
-        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillHide), name: UIResponder.keyboardWillHideNotification, object: nil)
-    }
-    
-    @objc private func keyboardWillShow(_ notification: Notification) {
-        guard let userInfo = notification.userInfo as NSDictionary?,
-              var keyboardFrame = (userInfo[UIResponder.keyboardFrameEndUserInfoKey] as? NSValue)?.cgRectValue else {
-            return
-        }
-        
-        var contentInset = textView.contentInset
-        keyboardFrame = view.convert(keyboardFrame, from: nil)
-        contentInset.bottom = keyboardFrame.size.height
-        textView.contentInset = contentInset
-        textView.verticalScrollIndicatorInsets = textView.contentInset
-    }
-    
-    @objc private func keyboardWillHide() {
-        textView.contentInset = UIEdgeInsets.zero
-        textView.verticalScrollIndicatorInsets = textView.contentInset
+    private func setUpKeyboard() {
+        keyboardManager = KeyboardManager(textView: textView)
     }
 }
